Rename mdns implementation, make it a module instead of package-level API (API change!)
