+++ Zeroconf/dns.py
@@ -1,10 +1,8 @@
 """ Multicast DNS Service Discovery for Python, v0.12
     Copyright (C) 2003, Paul Scott-Murphy
 
-    This module provides a framework for the use of DNS Service Discovery
-    using IP multicast.  It has been tested against the JRendezvous
-    implementation from <a href="http://strangeberry.com">StrangeBerry</a>,
-    and against the mDNSResponder from Mac OS X 10.3.8.
+    This module provides a DNS/mDNS encoding/decoding facility which
+    is used by the package to communicate with mDNS servers/clients.
 
     This library is free software; you can redistribute it and/or
     modify it under the terms of the GNU Lesser General Public
@@ -19,10 +17,6 @@
     You should have received a copy of the GNU Lesser General Public
     License along with this library; if not, write to the Free Software
     Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
-
-
-This module contains the mDNS/DNS server required to provide mDNS service
-discovery
 """
 import string
 import time
@@ -38,6 +32,7 @@ __all__ = [
     'DNSText',
     'AbstractMethodException', 'BadTypeInNameException',
     'NamePartTooLongException', 'NonLocalNameException', 'NonUniqueNameException',
+    'currentTimeMillis',
 ]
 
 # Some timing constants
