+++ samplecode/testmulticast.py
@@ -6,17 +6,17 @@ join the mDNS multicast group and send a (malformed) message to the group,
 our socket should receive that packet (because we have enabled multicast
 loopback on the socket).
 """
-import mcastsocket,socket,os,sys,select,logging
-import Zeroconf
+import socket,os,sys,select,logging
+from Zeroconf import dns,mcastsocket,mdns
 
 def main(ip):
     """Create a multicast socket, send a message, check it comes back"""
-    sock = mcastsocket.create_socket( (ip,Zeroconf._MDNS_PORT), loop=True )
-    mcastsocket.join_group( sock, Zeroconf._MDNS_ADDR )
+    sock = mcastsocket.create_socket( (ip,dns._MDNS_PORT), loop=True )
+    mcastsocket.join_group( sock, dns._MDNS_ADDR )
     try:
         payload = 'hello world'
         for i in range( 5 ):
-            sock.sendto( payload, 0, (Zeroconf._MDNS_ADDR, Zeroconf._MDNS_PORT))
+            sock.sendto( payload, 0, (dns._MDNS_ADDR, dns._MDNS_PORT))
             print 'Waiting for looped message receipt'
             rs,wr,xs = select.select( [sock],[],[], 1.0 )
             data,(addr,port) = sock.recvfrom( 200 )
@@ -26,7 +26,7 @@ def main(ip):
         print 'Failure: Looped message not received'
         return 1
     finally:
-        mcastsocket.leave_group( sock, Zeroconf._MDNS_ADDR )
+        mcastsocket.leave_group( sock, dns._MDNS_ADDR )
 
 if __name__ == "__main__":
     logging.basicConfig( level = logging.DEBUG )
