+++ setup.py
@@ -1,9 +1,10 @@
 #! /usr/bin/env python
 """Install Zeroconf.py using distutils"""
 from distutils.core import setup
+import os
 info = {}
 keys = [('__author__','author'),('__email__','author_email'),('__version__','version')]
-for line in open( 'Zeroconf.py' ):
+for line in open( os.path.join('Zeroconf','__init__.py') ):
     for key,inf in keys:
         if line.startswith( key ):
             info[inf] = line.strip().split('=')[1].strip().strip('"').strip("'")
@@ -15,7 +16,7 @@ if __name__ == "__main__":
         name='pyzeroconf',
         description='Python Zeroconf (mDNS) Library',
         url='http://digitaltorque.ca',
-        py_modules=['Zeroconf'],
+        packages=['zeroconf'],
         #scripts=['Browser.py'],
         classifiers=[
             'Development Status :: Production',
